requirements.txt
pytest
requests
pyyaml
allure-pytest
pytest-rerunfailures
pytest-xdist